make sure we test for both implementation of multitree the equals and hashcode functions
