+++ src/test/java/com/github/giedomak/telepathdb/datamodels/plans/AbstractMultiTreeTest.kt
@@ -2,10 +2,9 @@ package com.github.giedomak.telepathdb.datamodels.plans
 
 import com.github.giedomak.telepathdb.datamodels.graph.Edge
 import com.nhaarman.mockito_kotlin.mock
-import org.junit.Assert
 import org.junit.Test
 import java.util.stream.IntStream
-import kotlin.test.assertNotEquals
+import kotlin.test.assertEquals
 
 class AbstractMultiTreeTest {
 
@@ -31,38 +30,6 @@ class AbstractMultiTreeTest {
         val actual = root.postOrderTraversal().map { it.leaf!!.label }.toList().toString()
 
         // then
-        Assert.assertEquals(expected, actual)
-    }
-
-    @Test
-    fun equalsImplementation() {
-        val a = LogicalPlanTest.generateLogicalPlan(LogicalPlan.CONCATENATION, listOf("a", "b"))
-        val b = LogicalPlanTest.generateLogicalPlan(LogicalPlan.CONCATENATION, listOf("a", "b"))
-        val c = LogicalPlanTest.generateLogicalPlan(LogicalPlan.UNION, listOf("a", "b"))
-        val d = LogicalPlanTest.generateLogicalPlan(LogicalPlan.CONCATENATION, listOf("a", "b", "c"))
-        val e = LogicalPlanTest.generateLogicalPlan(LogicalPlan.UNION, listOf("a", "b", "c"))
-
-        Assert.assertEquals(a, b)
-        assertNotEquals(a, c)
-        assertNotEquals(a, d)
-        assertNotEquals(a, e)
-        assertNotEquals(c, e)
-        assertNotEquals(d, e)
-    }
-
-    @Test
-    fun generatesHashCode() {
-        val a = LogicalPlanTest.generateLogicalPlan(LogicalPlan.CONCATENATION, listOf("a", "b")).hashCode()
-        val b = LogicalPlanTest.generateLogicalPlan(LogicalPlan.CONCATENATION, listOf("a", "b")).hashCode()
-        val c = LogicalPlanTest.generateLogicalPlan(LogicalPlan.UNION, listOf("a", "b")).hashCode()
-        val d = LogicalPlanTest.generateLogicalPlan(LogicalPlan.CONCATENATION, listOf("a", "b", "c")).hashCode()
-        val e = LogicalPlanTest.generateLogicalPlan(LogicalPlan.UNION, listOf("a", "b", "c")).hashCode()
-
-        Assert.assertEquals(a, b)
-        assertNotEquals(a, c)
-        assertNotEquals(a, d)
-        assertNotEquals(a, e)
-        assertNotEquals(c, e)
-        assertNotEquals(d, e)
+        assertEquals(expected, actual)
     }
 }
